Test for job
A program I had to think of for an application for a job. It asked for me to not use for or while loops as well as i'm assuming switch because the application had bespoke programming language.

Writing it out helped me think about the problem, as the problem was building a factorial calculator without for loops.